# HTTP/1 headers simplification & cleanup