--- a/IntegrationTests/tests_01_http/test_07_headers_work.sh
+++ b/IntegrationTests/tests_01_http/test_07_headers_work.sh
@@ -19,7 +19,7 @@ token=$(create_token)
 start_server "$token"
 do_curl "$token" -H "foo: bar" --http1.0 \
     "http://foobar.com/dynamic/info" > "$tmp/out"
-if ! grep -q '("foo", "bar")' "$tmp/out"; then
+if ! grep -q '(name: "foo", value: "bar")' "$tmp/out"; then
     fail "couldn't find header in response"
 fi
 stop_server "$token"