--- a/Sources/NIOHTTP1/ByteCollectionUtils.swift
+++ b/Sources/NIOHTTP1/ByteCollectionUtils.swift
@@ -39,8 +39,35 @@ extension Sequence where Self.Element == UInt8 {
     ///
     /// - Parameter bytes: The string constant in the form of a collection of `UInt8`
     /// - Returns: Whether the collection contains **EXACTLY** this array or no, but by ignoring case.
-    internal func compareCaseInsensitiveASCIIBytes<T: Sequence>(to bytes: T) -> Bool
+    internal func compareCaseInsensitiveASCIIBytes<T: Sequence>(to: T) -> Bool
         where T.Element == UInt8 {
-            return self.elementsEqual(bytes, by: {return ($0 & 0xdf) == ($1 & 0xdf)})
+            // fast path: we can get the underlying bytes of both
+            let maybeMaybeResult = self.withContiguousStorageIfAvailable { lhsBuffer -> Bool? in
+                to.withContiguousStorageIfAvailable { rhsBuffer in
+                    if lhsBuffer.count != rhsBuffer.count {
+                        return false
+                    }
+
+                    for idx in 0 ..< lhsBuffer.count {
+                        // let's hope this gets vectorised ;)
+                        if lhsBuffer[idx] & 0xdf != rhsBuffer[idx] & 0xdf {
+                            return false
+                        }
+                    }
+                    return true
+                }
+            }
+
+            if let maybeResult = maybeMaybeResult, let result = maybeResult {
+                return result
+            } else {
+                return self.elementsEqual(to, by: {return ($0 & 0xdf) == ($1 & 0xdf)})
+            }
+    }
+}
+
+extension String {
+    internal func isEqualCaseInsensitiveASCIIBytes(to: String) -> Bool {
+        return self.utf8.compareCaseInsensitiveASCIIBytes(to: to.utf8)
     }
 }