--- a/docker/docker-compose.1804.50.yaml
+++ b/docker/docker-compose.1804.50.yaml
@@ -19,7 +19,7 @@ services:
   integration-tests:
     image: swift-nio:18.04-5.0
     environment:
-      - MAX_ALLOCS_ALLOWED_1000_reqs_1_conn=45750
+      - MAX_ALLOCS_ALLOWED_1000_reqs_1_conn=31200
       - MAX_ALLOCS_ALLOWED_1_reqs_1000_conn=1177050 # was: 685050
       - MAX_ALLOCS_ALLOWED_ping_pong_1000_reqs_1_conn=4600
       - MAX_ALLOCS_ALLOWED_bytebuffer_lots_of_rw=2100
@@ -29,7 +29,7 @@ services:
     image: swift-nio:18.04-5.0
     command: /bin/bash -cl "swift test -Xswiftc -warnings-as-errors -Xswiftc -DNIO_CI_BUILD && ./scripts/integration_tests.sh"
     environment:
-      - MAX_ALLOCS_ALLOWED_1000_reqs_1_conn=45750
+      - MAX_ALLOCS_ALLOWED_1000_reqs_1_conn=31200
       - MAX_ALLOCS_ALLOWED_1_reqs_1000_conn=1177050 # was: 685050
       - MAX_ALLOCS_ALLOWED_ping_pong_1000_reqs_1_conn=4600
       - MAX_ALLOCS_ALLOWED_bytebuffer_lots_of_rw=2100